backend/requirements.txt
Flask==3.1.2
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.0
Werkzeug==3.1.3
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.2.0
click==8.1.7
bcrypt==4.3.0
passlib==1.7.4
SQLAlchemy-Utils==0.41.1

Flask-JWT-Extended==4.6.0
flask-injector
psycopg2-binary==2.9.7
requests==2.31.0
gunicorn==21.2.0
PyJWT==2.8.0
pytz==2024.1
cryptography==41.0.3
numpy==1.24.3
pandas==2.0.3
pytest==7.4.0
pytest-mock==3.12.0Flask-Migrate
